requirements.txt
aiohttp
python-dotenv
asyncio
pydantic
uvicorn
fastapi